Fixing float swap, and everything that uses it

+++ src/utils.c
@@ -7,7 +7,7 @@ void int_swap(int *a, int *b) {
 }
 
 void float_swap(float *a, float *b) {
-    int temp = *a;
+    float temp = *a;
     *a = *b;
     *b = temp;
 }
